#4 feat: 카테고리,키워드 가져오기 api 연동 완료
