+++ src/components/mainPage/Search.tsx
@@ -14,6 +14,7 @@ import {
   SearchText,
   Si,
 } from "../../store/searchAtom";
+import KeyWordInput from "./KeyWordInput";
 
 const Search = () => {
   const navigate = useNavigate();
@@ -29,11 +30,6 @@ const Search = () => {
   const [keyWord, setKeyWord] = useRecoilState<string>(KeyWord);
   const [pageNum, setPageNum] = useRecoilState<string>(PageNum);
 
-  //DB에 저장된 검색가능한 키워드 리스트 가져오기
-  const { data: keyWordList } = useQuery(["keyWordList"], getKeyWordList, {
-    refetchOnWindowFocus: false,
-  });
-
   //onClick
   const categoriesBtnOnClick = () => {
     setCategoriesBoolean(true);
@@ -42,10 +38,6 @@ const Search = () => {
     if (city === "전체 지역") setCity("서울");
     setCategoriesBoolean(false);
     setSi("종로구");
-  };
-  const keyWordOnClick = (value: string) => {
-    if (value === keyWord) setKeyWord("");
-    else setKeyWord(value);
   };
 
   //onChange
@@ -114,21 +106,7 @@ const Search = () => {
             </SearchInputBtnBox>
           </SearchBox>
         </DateAndSearchBox>
-        <KeyWordBox>
-          <KeyWordList>
-            {Array.isArray(keyWordList)
-              ? keyWordList.map((value: any, index: any) => (
-                  <KeyWordItem
-                    active={keyWord === value.name}
-                    onClick={() => keyWordOnClick(value.name)}
-                    key={value.keywordId}
-                  >
-                    # {value.name}
-                  </KeyWordItem>
-                ))
-              : null}
-          </KeyWordList>
-        </KeyWordBox>
+        <KeyWordInput keyWord={keyWord} setKeyWord={setKeyWord} />
       </SearchContainer>
     </>
   );
@@ -227,50 +205,6 @@ const SearchBtn = styled.button`
   color: ${({ theme }) => theme.colors.white};
   font-size: 1rem;
   cursor: pointer;
-`;
-
-const KeyWordBox = styled.div`
-  overflow: hidden;
-`;
-const KeyWordList = styled.ul`
-  list-style: none;
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  @media (max-width: 680px) {
-    justify-content: normal;
-    padding-bottom: 5px;
-    overflow-x: scroll;
-    &::-webkit-scrollbar {
-      width: 3px;
-      height: 7px;
-    }
-    &::-webkit-scrollbar-thumb {
-      width: 3px;
-      border-radius: 5px;
-      background-color: #b1b1b1;
-    }
-    &::-webkit-scrollbar-track {
-      background-color: #ffffff;
-    }
-  }
-`;
-const KeyWordItem = styled.li<{ active: boolean }>`
-  margin: 0 5px;
-  padding: 5px 10px;
-  min-width: fit-content;
-  height: 100%;
-  background-color: #d9d9d9;
-  border-radius: 10px;
-  text-align: center;
-  font-size: 0.7rem;
-  cursor: pointer;
-  ${(props) =>
-    props.active &&
-    css`
-      background-color: #0075ff;
-      color: ${({ theme }) => theme.colors.white};
-    `}
 `;
 
 const CategoriesBtn = styled.label`
